kategori

+++ application/controllers/dashboard.php
@@ -1,18 +1,19 @@
 <?php
 
-class Dashboard extends CI_Controller
-{
+class Dashboard extends CI_Controller{
 
-    public
-     function index()
-     {
-        $data['barang'] = $this->model_barang->tampil_data()->result();
-        $this->load->view('templates/header');
-        $this->load->view('templates/sidebar');
-        $this->load->view('dashboard', $data);
-        $this->load->view('templates/footer');
+   public function __construct(){
+      parent::__construct();
 
-     }
+      if($this->session->userdata('role_id') != '2'
+      ){
+          $this->session->set_flashdata('pesan','<div class="alert alert-danger alert-dismissible fade show" role="alert">
+          Anda Belum Login!
+          <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
+        </div>');
+        redirect('auth/login');
+      }
+  }
 
      public function tambah_ke_keranjang($id)
      {
@@ -27,7 +28,7 @@ class Dashboard extends CI_Controller
          );
 
          $this->cart->insert($data);
-         redirect('dashboard');
+         redirect('welcome');
      }
 
      public function detail_keranjang()
@@ -41,7 +42,7 @@ class Dashboard extends CI_Controller
      public function hapus_keranjang()
      {
       $this->cart->destroy();
-      redirect('dashboard/index');
+      redirect('welcome/index');
      }
 
      public function pembayaran()
@@ -54,10 +55,25 @@ class Dashboard extends CI_Controller
 
      public function proses_pesanan()
      {
-      $this->cart->destroy();
+         $is_processed = $this->model_invoice->index();
+         if($is_processed){
+            $this->cart->destroy();
+            $this->load->view('templates/header');
+            $this->load->view('templates/sidebar');
+            $this->load->view('proses_pesanan');
+            $this->load->view('templates/footer');
+         } else {
+            echo "Maaf, Pesanan Anda Gagal Di Proses!";
+         }
+     }
+
+     public function detail($id_brg)
+     {
+      $data['barang'] = $this->model_barang->detail_brg($id_brg);
       $this->load->view('templates/header');
       $this->load->view('templates/sidebar');
-      $this->load->view('proses_pesanan');
+      $this->load->view('detail_barang', $data);
       $this->load->view('templates/footer');
      }
+
 }
